Discipulado is now saved into the DB

+++ src/main/java/com/example/database_user/services/DiscipuladoMenoresService.java
@@ -118,6 +118,10 @@ public class DiscipuladoMenoresService {
      */
     public ResponseEntity<String> updateDiscipuladoMenores(DiscipuladoMenores discipuladoMenores) {
 
+        if (discipuladoMenores.getId() == null) {
+            this.insertNewDiscipuladoMenores(discipuladoMenores);
+            return new ResponseEntity<>(HttpStatus.CREATED);
+        }
         discipuladoMenoresRepository.deleteById(discipuladoMenores.getId());
         discipuladoMenoresRepository.insert(discipuladoMenores);
         return new ResponseEntity<>(HttpStatus.OK);
